+++ src/bloom_filter/bloom_filter_calculate_num_hashes_shadow.c
@@ -25,7 +25,7 @@ unsigned int nondet_uint();
  * \return The number of hash functions required.
  */
 unsigned int bloom_filter_calculate_num_hashes(
-    unsigned int num_expected_entries, size_t size_in_bytes)
+    size_t num_expected_entries, size_t size_in_bytes)
 {
     unsigned int n = nondet_uint();
